Extract element resize detection from hook
